Move some methods to point mutationgo
